+++ client/oslc-client/src/main/java/org/eclipse/lyo/client/OslcOAuthClientBuilder.java
@@ -13,11 +13,10 @@
  */
 package org.eclipse.lyo.client;
 
-import javax.ws.rs.client.ClientBuilder;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import jakarta.ws.rs.client.ClientBuilder;
 import net.oauth.OAuthAccessor;
 import net.oauth.OAuthConsumer;
 import net.oauth.OAuthServiceProvider;